refactor: move fetch users to users store
